+++ package.json
@@ -21,9 +21,10 @@
   "homepage": "https://github.com/skycrazyk/enchat#readme",
   "dependencies": {
     "@reduxjs/toolkit": "^1.3.6",
+    "@types/mini-css-extract-plugin": "^0.9.1",
     "antd": "^4.4.1",
-    "lodash": "^4.17.15",
     "fastclick": "^1.0.6",
+    "lodash": "^4.17.15",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-hot-loader": "^4.12.21",
@@ -51,6 +52,7 @@
     "cross-env": "^7.0.2",
     "css-loader": "^3.5.3",
     "html-webpack-plugin": "^4.3.0",
+    "mini-css-extract-plugin": "^0.9.0",
     "postcss-loader": "^3.0.0",
     "prettier": "^2.0.5",
     "style-loader": "^1.2.1",
